Set the time zone to UTC for PostgreSQL connections.
This mitigates an issue with the timestamp type which was exposed by the
addition of type information for the ptime field.  Conversions from
timestamp with time zone to timestamp converts the time to local time.
Earlier versions worked since the time zone is merely stripped off when
converting a string to a timestamp.