Disallow ?i non-linear substitutions
In the discussion for PR #77 @paurkedal brought up that caqti offers a
choice between `?` and `$n`. In practice, it was possible to use `?n`
with the sqlite3 driver as long as the number of substitutions
corresponds to the number of arguments. The following query would work:

    SELECT * FROM house WHERE color = ?2 AND city = ?1

while the following would fail:

    SELECT * FROM house WHERE color = ?2 AND (?1 IS NULL OR city = ?1)