Add a printf-like `Caqti_query.qprintf` for dynamic queries
Sometimes, you need to construct queries dynamically rather than writing
them out in code. Building an ORM would be an obvious reason for doing
that, but it is quite often that one has to sprinkle a little bit of
dynamism here and there.

In the case that one has to generate queries dynamically, the most
obvious way coming from standard Caqti usage is to build up a query
string fragment using Format or Printf, then use it as if you had
written it manually. Simple. But then what happens is that Caqti will
take the string that you have built dynamically, and parse it back into
a `Caqti_query.t` before using the fragments to build the actual query
string.

That doesn't sound very efficient, and relies on you (as the
dynamic-query-generator) not messing up the generation of things like
quoted strings if you don't want to expose security holes in your
software.  In fact, Caqti's documentation quietly recommends using the
`Caqti_query.t` type directly when building dynamic queries, touting its
easy composability.

While true (`Caqti_query.t` is easily composable), it loses out on
convenience compared to the Format + parse approach. For instance, if
you want to generate, say, the MAX of two dynamically generated queries,
you have to write `S [ L "MAX("; q1; L ", "; q2; L ")" ]`. Maybe if you
are keen on combinators, you can figure out how to write a somewhat more
readable version such as `l "MAX(" %% q1 % ", " %% q2 % ")"`. But you
never get something as nice as the `"MAX(%a, %a)"` that you get with
Format.

Or at least, that's what I thought before realizing that format6 is
actually *very* polymorphic. It has to, so that it can accomodate the
whole family of `*printf` functions. And it turns out that it is
polymorphic enough to write a `printf`-style function to build a
`Caqti_query.t`.

So here is `qprintf`. Like `sprintf`, but builds a `Caqti_query.t`
instead of a `string`. You can delimit quotes and environment variables
using tags (`@{<Q>the quote@}` and `@{<E>env.var@}`, respectively), and
include parameters with `param`.

This gives a convenient, readable and secure (w.r.t. quoting) way of
building queries dynamically, which I thought was worth sharing.