Turn deprecations back into alerts for type constructor aliases.
This reverts the previous commit except for the added descriptions.
There is no stable alternative at the moment, to the best option is for
anyone using the constructors to keep the current code, disable the
private warning, which will be needed also in the future, and use a
tight version constraint on Caqti.